Change modules view for user by custom reexport. Improved documentation.
